+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added parameter for the retain property for Send-MQTTMessage (-Retain) (defaults to false)
+- Added parameter for the qoslevel property for Send-MQTTMessage (-Qoslevel) (defaults to 0)
+
 ## [1.1.0] - 2024-03-27
 
 ### Added